Buffer only requested data in VCD loader
Previously, VcdLoader was reading in the whole file on construction. This
is prohibitive even for relatively small files of just a few megabyte.
Now, we still walk through the file from the beginning, but we only record
signals and cycles that are requested, i.e. currently on the screen. This
means memory footprint is roughly constant independent of file size. We
pay for this in performance, because the file has to be re-read when screen
contents changes.